A cast returns what it found; the command decides what to say about it
`castMold(request)` is the whole assembly, and it reads nothing it was not
handed. Where this repo keeps its Molds, its targets and its reference contract
are facts about this tree, so the command finds them and passes what it found —
a request carrying the loaded target, the parsed Mold, the corpus index, the
hooks, and the block this Foundry records beside `refs`.

It also prints nothing. Errors and drift come back as values, because a cast
that found four unresolved refs has produced a result rather than suffered a
failure, and how that gets rendered is the caller's question. `reportCast` is
the only thing left that knows about stderr and exit codes.

The one decision that stays inside is whether to publish, since that is a
question about the staged bundle. `wrote` says what happened; the reporter reads
it rather than re-deriving the rule from `errors`, so the two cannot disagree.

`cast_target` now comes from the target the caller asked for rather than
`_target.yml`'s `name:` restating it. Same string in this repo — a declaration
that disagreed with the directory holding it is worth reporting, not worth
quietly writing into every record.

Oracle clean; a regenerate moves 94 lines over 47 bundles, `cast_at` and
`commit` and nothing else; 299 tests pass.
